refactor(auth): move auth dtos to auth/dtos directory and update imports
Restructure auth-related DTOs into a dedicated auth/dtos directory for better organization and maintainability. Update all import paths to reflect the new location. Includes moving test files and updating test imports accordingly.
