+++ src/config/__tests__/__helpers__/auth-datasource.helpers.ts
@@ -1,6 +1,6 @@
-import { RegisterDto } from "@/domain/dtos/register.dto";
-import { LoginDto } from "@/domain/dtos/login.dto";
-import { LogoutDto } from "@/domain/dtos/logout.dto";
+import { RegisterDto } from "@/domain/auth/dtos/register.dto";
+import { LoginDto } from "@/domain/auth/dtos/login.dto";
+import { LogoutDto } from "@/domain/auth/dtos/logout.dto";
 import { UpdateUserDto } from "@/domain/dtos/update-user.dto";
 import { UserEntity } from "@/domain/entities/user.entity";
 import { AuthUserEntity } from "@/domain/entities/auth-user.entity";
