+++ src/domain/auth/dtos/__tests__/register.dto.test.ts
