+++ src/domain/auth/dtos/logout.dto.ts
@@ -1,4 +1,4 @@
-import { TRawJson } from "../interfaces/general.interfaces";
+import { TRawJson } from "../../interfaces/general.interfaces";
 import { ERRORS } from "@/config/strings/global.strings.json";
 
 export class LogoutDto {
