+++ src/domain/use-cases/__tests__/login-user.test.ts
@@ -1,5 +1,5 @@
 import { LoginUser } from "../login-user";
-import { LoginDto } from "@/domain/dtos/login.dto";
+import { LoginDto } from "@/domain/auth/dtos/login.dto";
 import { AuthUserEntity } from "@/domain/entities/auth-user.entity";
 import { UserEntity } from "@/domain/entities/user.entity";
 import { MockAuthRepository } from "../../../config/__tests__/mocks/auth-repository.mock";
