+++ src/domain/use-cases/__tests__/logout-user.test.ts
@@ -1,5 +1,5 @@
 import { LogoutAuth } from "../logout-user";
-import { LogoutDto } from "@/domain/dtos/logout.dto";
+import { LogoutDto } from "@/domain/auth/dtos/logout.dto";
 import { MockAuthRepository } from "../../../config/__tests__/mocks/auth-repository.mock";
 
 describe("LogoutAuth", () => {
