+++ src/domain/use-cases/__tests__/register-user.test.ts
@@ -1,5 +1,5 @@
 import { RegisterUser } from "../register-user";
-import { RegisterDto } from "../../dtos/register.dto";
+import { RegisterDto } from "../../auth/dtos/register.dto";
 import { AuthUserEntity } from "../../entities/auth-user.entity";
 import { UserEntity } from "../../entities/user.entity";
 import { MockAuthRepository } from "../../../config/__tests__/mocks/auth-repository.mock";
