+++ src/domain/use-cases/__tests__/request-reset-password-email.test.ts
@@ -1,5 +1,5 @@
 import { RequestResetPasswordEmail } from "../request-reset-password-email";
-import { RequestResetPasswordEmailDto } from "../../dtos/request-reset-password-email.dto";
+import { RequestResetPasswordEmailDto } from "../../auth/dtos/request-reset-password-email.dto";
 import { MockAuthRepository } from "../../../config/__tests__/mocks/auth-repository.mock";
 
 describe("RequestResetPasswordEmail", () => {
