+++ src/domain/use-cases/register-user.ts
@@ -1,4 +1,4 @@
-import { RegisterDto } from "../dtos/register.dto";
+import { RegisterDto } from "../auth/dtos/register.dto";
 import { AuthUserEntity } from "../entities/auth-user.entity";
 import { UserEntity } from "../entities/user.entity";
 import { AuthRepository } from "../repositories/auth.repository";
