+++ src/infrastructure/datasources/__tests__/logout.test.ts
@@ -1,6 +1,6 @@
 import { AuthDatasource } from "../auth.datasource";
 import { AuthClient } from "@/infrastructure/config/auth.client";
-import { LogoutDto } from "@/domain/dtos/logout.dto";
+import { LogoutDto } from "@/domain/auth/dtos/logout.dto";
 import { createMockLogoutDto } from "@/config/__tests__/__helpers__/auth-datasource.helpers";
 
 // Mock dependencies
