+++ src/infrastructure/datasources/auth.datasource.ts
@@ -1,4 +1,4 @@
-import { RegisterDto } from "@/domain/dtos/register.dto";
+import { RegisterDto } from "@/domain/auth/dtos/register.dto";
 import { AuthUserEntity } from "@/domain/entities/auth-user.entity";
 import { UserEntity } from "@/domain/entities/user.entity";
 import { AuthRepository } from "@/domain/repositories/auth.repository";
@@ -7,10 +7,10 @@ import { ERRORS } from "@/config/strings/global.strings.json";
 import { DatasourceUserDto } from "../dtos/datasource-user.dto";
 import { BadRequestError } from "@/domain/errors/bad-request-error";
 import { ValidationError } from "@/domain/errors/validation-error";
-import { LoginDto } from "@/domain/dtos/login.dto";
-import { LogoutDto } from "@/domain/dtos/logout.dto";
+import { LoginDto } from "@/domain/auth/dtos/login.dto";
+import { LogoutDto } from "@/domain/auth/dtos/logout.dto";
 import { UpdateUserDto } from "@/domain/dtos/update-user.dto";
-import { RequestResetPasswordEmailDto } from "@/domain/dtos/request-reset-password-email.dto";
+import { RequestResetPasswordEmailDto } from "@/domain/auth/dtos/request-reset-password-email.dto";
 import { SupabaseClient } from "@supabase/supabase-js";
 import { TRawJson } from "@/domain/interfaces/general.interfaces";
